requirements.txt
fastapi>=0.116.0
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.1
google-genai>=1.25.0
PyGithub==1.59.1
requests==2.31.0
python-dotenv==1.0.0
openai>=1.0.0
beautifulsoup4>=4.12.0
mem0ai>=0.1.114
